requirements.txt
# Core Flask + Extensions
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-Limiter==3.5.1
Flask-Talisman==1.1.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1

# Server / WebSockets
gunicorn==21.2.0
eventlet==0.35.1  # Required for Flask-SocketIO on Render

# Async Task Support (Optional if using SocketIO with background tasks)
redis==5.0.4

# Security & Utilities
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
requests==2.32.3

# Logging
loguru==0.7.2

# Email
secure-smtplib==0.1.1  # Optional if using secure SMTP

# Image Processing
opencv-python==4.9.0.80
mediapipe==0.10.9
deepface==0.40.1

# Audio (Optional for TTS/Speech Features)
playsound==1.2.2
SpeechRecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.13
gTTS==2.5.1
